+++ R/ops.R
@@ -1,6 +1,6 @@
 .ops_ts = \(data, target, source, E = 3:5, k = E, tau = 1, style = 1, lib = NULL, pred = NULL,
             maximize = c("dark", "positive", "negative"), dist.metric = c("euclidean", "manhattan", "maximum"), 
-            zero.tolerance = max(k), relative = TRUE, weighted = TRUE, threads = length(E), higher.parallel = TRUE, h = 0) {
+            zero.tolerance = max(k), relative = TRUE, weighted = TRUE, threads = length(E), higher.parallel = TRUE, h = 0, ...) {
   maximize = match.arg(maximize)
   dist.metric = match.arg(dist.metric)
   dlist = .validate_var(data, target, source)
@@ -14,7 +14,7 @@
 
 .ops_lattice = \(data, target, source, E = 3:5, k = E+1, tau = 1, style = 1, lib = NULL, pred = NULL, 
                  maximize = c("dark", "positive", "negative"), dist.metric = c("euclidean", "manhattan", "maximum"), 
-                 zero.tolerance = max(k), relative = TRUE, weighted = TRUE, threads = length(E), higher.parallel = TRUE, detrend = FALSE, nb = NULL) {
+                 zero.tolerance = max(k), relative = TRUE, weighted = TRUE, threads = length(E), higher.parallel = TRUE, detrend = FALSE, nb = NULL, ...) {
   if (is.null(nb)) nb = sdsfun::spdep_nb(data)
   maximize = match.arg(maximize)
   dist.metric = match.arg(dist.metric)
@@ -29,7 +29,7 @@
 
 .ops_grid = \(data, target, source, E = 3:5, k = E+1, tau = 1, style = 1, lib = NULL, pred = NULL,
               maximize = c("positive", "negative", "dark"), dist.metric = c("euclidean", "manhattan", "maximum"), 
-              zero.tolerance = max(k), relative = TRUE, weighted = TRUE, threads = length(E), higher.parallel = TRUE, detrend = FALSE) {
+              zero.tolerance = max(k), relative = TRUE, weighted = TRUE, threads = length(E), higher.parallel = TRUE, detrend = FALSE, ...) {
   maximize = match.arg(maximize)
   dist.metric = match.arg(dist.metric)
   dlist = .validate_var(data, target, source, detrend)